% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{pilot3utils-package}
\alias{pilot3utils}
\alias{pilot3utils-package}
\title{pilot3utils: Utilities for the Pilot 3 Submission to the FDA}
\description{
R Consortium R Submission Pilot 3.
}
\author{
\strong{Maintainer}: Benjamin Straub-GSK \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Joel Laxamana-Roche
}

}
\keyword{internal}
